address (hopefully) circular imports issue; move dummy_fitstbl to tstutils.py
